PHRAS-3445 add edit retry and failed Q
