+++ .env
@@ -139,13 +139,13 @@ PHRASEANET_EXPLODE_WORKER=1
 PHRASEANET_WORKER_assetsIngest=1
 PHRASEANET_WORKER_createRecord=2
 PHRASEANET_WORKER_deleteRecord=2
+PHRASEANET_WORKER_editRecord=2
 PHRASEANET_WORKER_exportMail=2
 PHRASEANET_WORKER_exposeUpload=2
 PHRASEANET_WORKER_ftp=1
 PHRASEANET_WORKER_mainQueue=3
 PHRASEANET_WORKER_populateIndex=1
 PHRASEANET_WORKER_pullAssets=1
-PHRASEANET_WORKER_recordEdit=2
 PHRASEANET_WORKER_subdefCreation=1
 PHRASEANET_WORKER_subtitle=1
 PHRASEANET_WORKER_validationReminder=1
