+++ docker-compose.yml
@@ -156,13 +156,13 @@ services:
     - PHRASEANET_WORKER_assetsIngest
     - PHRASEANET_WORKER_createRecord
     - PHRASEANET_WORKER_deleteRecord
+    - PHRASEANET_WORKER_editRecord
     - PHRASEANET_WORKER_exportMail
     - PHRASEANET_WORKER_exposeUpload
     - PHRASEANET_WORKER_ftp
     - PHRASEANET_WORKER_mainQueue
     - PHRASEANET_WORKER_populateIndex
     - PHRASEANET_WORKER_pullAssets
-    - PHRASEANET_WORKER_recordEdit
     - PHRASEANET_WORKER_subdefCreation
     - PHRASEANET_WORKER_subtitle
     - PHRASEANET_WORKER_validationReminder
