+++ lib/Alchemy/Phrasea/WorkerManager/Queue/AMQPConnection.php
@@ -93,9 +93,10 @@ class AMQPConnection
             self::MAX_RETRY   => self::DEFAULT_MAX_RETRY_VALUE,
             self::TTL_RETRY   => self::DEFAULT_RETRY_DELAY_VALUE,
         ],
-        MessagePublisher::RECORD_EDIT_TYPE       => [
-            'with'            => self::WITH_NOTHING,
+        MessagePublisher::EDIT_RECORD_TYPE       => [
+            'with'            => self::WITH_RETRY,
             self::MAX_RETRY   => self::DEFAULT_MAX_RETRY_VALUE,
+            self::TTL_RETRY   => self::DEFAULT_RETRY_DELAY_VALUE,
         ],
         MessagePublisher::SUBDEF_CREATION_TYPE     => [
             'with'            => self::WITH_RETRY | self::WITH_DELAYED,
