+++ lib/Alchemy/Phrasea/WorkerManager/Queue/MessageHandler.php
@@ -127,6 +127,6 @@ class MessageHandler
 
         //  give prefetch message to a worker consumer at a time
         $channel->basic_qos(null, $prefetchCount, null);
-        $channel->basic_consume($queueName, Uuid::uuid4(), false, false, false, false, $callback);
+        $channel->basic_consume($queueName, Uuid::uuid4()->toString(), false, false, false, false, $callback);
     }
 }
